Implement confirmation method in BasicAuthController and associated tests
The confirmation method has been added to BasicAuthController to handle specific requests. This required the creation of new DTOs for the request and response as well as an update to BasicAuthService. The previous AuthControllerTest was deleted and a new WebLayerTest was created to test the updated functionality. The build.gradle.kts file was also updated to include a new dependency on Apache Commons Lang.