Add JavaDoc for methods in JwtService.
In the JwtService class, documentation comments were added for two methods. These additional JavaDocs clarify the usage of the methods which are responsible for creating a JSON Web Token (JWT) and retrieving the subject from a JWT. This enhancement in comments aids developers with better context and understanding of these methods' roles and their expected parameters and return values.