Enable deleteResumableUpload feature and add exception on failure
This commit updates the code to enable the deleteResumableUpload feature which was previously commented out in both FilesController and FilesService. Simultaneously, a new function to fetch a specific ResumableUpload based on id, as well as a function to delete related files, have been added to the Resumables class. Further, a new exception called FailedResourceDeletionException was introduced to handle potential failures during deletion.